+++ packages/components/atoms/f-image-tile/src/components/ImageTile.vue
@@ -10,7 +10,7 @@
                 $style['c-imageTile-link'], {
                     [$style['c-imageTile-link--toggle']]: !isLink
                 }]"
-            :href="isLink ? href : '#'"
+            :href="href"
             :aria-hidden="!isLink"
             :tabindex="isLink ? false : -1"
             data-test-id="image-tile-link">
@@ -21,6 +21,7 @@
         <input
             :id="`imageTileToggle-${tileId}`"
             type="checkbox"
+            :checked="isToggleSelected"
             class="is-visuallyHidden"
             :class="$style['c-imageTile-checkbox']"
             data-test-id="image-tile-input"
@@ -50,7 +51,9 @@
                 :class="$style['c-imageTile-textContainer']"
                 :aria-hidden="isLink">
                 <tick-icon :class="$style['c-imageTile-icon']" />
-                <span :class="$style['c-imageTile-text']">
+                <span
+                    :class="$style['c-imageTile-text']"
+                    data-test-id="image-tile-text">
                     {{ displayText }}
                 </span>
             </span>
@@ -182,8 +185,8 @@ $image-tile-text-transform: translate3d(5px, 0, 0);
 }
 
 @mixin image-tile-focus() {
-    outline: 2px solid $color-focus;
     border-radius: $radius-rounded-b;
+    box-shadow: 0 0 0 2px $color-focus;
 }
 
 .c-imageTile {