+++ packages/components/atoms/f-image-tile/src/components/_tests/ImageTile.test.js
@@ -14,9 +14,9 @@ describe('ImageTile', () => {
             const isLinkFalse = false;
 
             it.each([
-                ['https://www.google.com', isLinkTrue],
-                ['#', isLinkFalse]
-            ])('should update `href` to %s when set to %s', (expectedValue, isLink) => {
+                ['https://www.google.com', undefined, isLinkTrue],
+                ['https://www.google.com', '-1', isLinkFalse]
+            ])('should pass `href` and update `tabindex` to expected values', (expectedValue, expectedTabindex, isLink) => {
                 // Arrange
                 const propsData = { isLink, href: 'https://www.google.com' };
 
@@ -29,6 +29,58 @@ describe('ImageTile', () => {
 
                 // Assert
                 expect(link.attributes('href')).toBe(expectedValue);
+                expect(link.attributes('tabindex')).toBe(expectedTabindex);
+            });
+        });
+
+        describe('href :: ', () => {
+            it('should apply `href` to the link', () => {
+                // Arrange
+                const propsData = { href: 'https://www.google.com' };
+
+                // Act
+                const wrapper = shallowMount(ImageTile, {
+                    propsData
+                });
+
+                const link = wrapper.find('[data-test-id="image-tile-link"]');
+
+                // Assert
+                expect(link.attributes('href')).toBe(propsData.href);
+            });
+        });
+
+        describe('displayText :: ', () => {
+            it('should apply `displayText` to the image tile', () => {
+                // Arrange
+                const propsData = { displayText: 'Chicken' };
+
+                // Act
+                const wrapper = shallowMount(ImageTile, {
+                    propsData
+                });
+
+                const text = wrapper.find('[data-test-id="image-tile-text"]');
+
+                // Assert
+                expect(text.text()).toBe(propsData.displayText);
+            });
+        });
+
+        describe('altText :: ', () => {
+            it('should apply `altText` to the image alt', () => {
+                // Arrange
+                const propsData = { imgSrc: 'https://via.placeholder.com/150', altText: '' };
+
+                // Act
+                const wrapper = shallowMount(ImageTile, {
+                    propsData
+                });
+
+                const image = wrapper.find('[data-test-id="image-tile-image"]');
+
+                // Assert
+                expect(image.attributes('alt')).toContain(propsData.altText);
             });
         });