+++ packages/components/organisms/f-cookie-banner/test-utils/component-objects/f-cookie-banner-selectors.js
@@ -1,0 +1,5 @@
+exports.COOKIE_BANNER_COMPONENT = '[data-test-id="cookieConsentBanner"]';
+exports.COOKIE_BANNER_TITLE = '[data-test-id="cookieConsentTitle"]';
+exports.COOKIE_POLICY_LINK = '[data-test-id="link-component"]';
+exports.COOKIE_ACCEPT_ALL_BUTTON = '[data-test-id="accept-all-cookies-button"]';
+exports.COOKIE_ACCEPT_NECESSARY_BUTTON = '[data-test-id="accept-necessary-cookies-button"]';