+++ packages/components/organisms/f-cookie-banner/test-utils/component-objects/f-cookie-consent-banner.component.js
@@ -1,19 +1,28 @@
 const Page = require('@justeat/f-wdio-utils/src/page.object');
+const {
+    COOKIE_BANNER_COMPONENT,
+    COOKIE_BANNER_TITLE,
+    COOKIE_POLICY_LINK,
+    COOKIE_ACCEPT_ALL_BUTTON,
+    COOKIE_ACCEPT_NECESSARY_BUTTON
+} = require('./f-cookie-banner-selectors');
 
 module.exports = class CookieBanner extends Page {
     constructor () {
         super('organism', 'cookie-banner-component');
     }
 
-    get component () { return $('[data-test-id="cookieConsentBanner"]'); }
+    get component () { return $(COOKIE_BANNER_COMPONENT); }
 
-    get cookiePolicyLink () { return this.component.$('[data-test-id="cookie-policy-link"]'); }
+    get cookieBannerTitle () { return $(COOKIE_BANNER_TITLE); }
+
+    get cookiePolicyLink () { return $(COOKIE_POLICY_LINK); }
 
     get cookiePolicyLinkHref () { return this.cookiePolicyLink.getAttribute('href'); }
 
-    get cookieAcceptAllButton () { return this.component.$('[data-test-id="accept-all-cookies-button"]'); }
+    get cookieAcceptAllButton () { return $(COOKIE_ACCEPT_ALL_BUTTON); }
 
-    get cookieAcceptNecessaryButton () { return this.component.$('[data-test-id="accept-necessary-cookies-button"]'); }
+    get cookieAcceptNecessaryButton () { return $(COOKIE_ACCEPT_NECESSARY_BUTTON); }
 
     load () {
         this.open('/');
@@ -44,5 +53,17 @@ module.exports = class CookieBanner extends Page {
         };
 
         return cookieType[value].click();
+    }
+
+    testTabOrder (tabOrder) {
+        const tabOrderResult = super.testTabOrder(tabOrder);
+        const expectedTabOrder = tabOrder.map(el => ({
+            selector: el.getAttribute('data-test-id'),
+            isFocused: true
+        }));
+        return {
+            actual: tabOrderResult,
+            expected: expectedTabOrder.concat(expectedTabOrder[0])
+        };
     }
 };