+++ packages/components/pages/f-account-info/src/components/_tests/AccountInfo.test.js
@@ -1,8 +1,6 @@
 import Vuex from 'vuex';
 import { shallowMount } from '@vue/test-utils';
 import AccountInfo from '../AccountInfo.vue';
-// eslint-disable-next-line no-unused-vars
-import AccountInfoAnalyticsService from '../../services/analytics';
 import AccountInfoValidationMixin from '../AccountInfoValidationMixin.vue';
 
 import {
@@ -16,12 +14,11 @@ import {
 let wrapper;
 let cookiesSpy;
 let httpSpy;
+let pushEventSpy;
 let sutMocks;
 let sutProps;
 let dataDefaults;
 let initialiseSpy;
-
-jest.mock('../../services/analytics.js');
 
 const logMocks = {
     info: jest.fn(),
@@ -30,7 +27,8 @@ const logMocks = {
 
 const storeActions = {
     loadConsumerDetails: jest.fn(),
-    editConsumerDetails: jest.fn()
+    editConsumerDetails: jest.fn(),
+    saveConsumerDetails: jest.fn()
 };
 
 const storeState = consumerStateModel;
@@ -76,17 +74,21 @@ describe('AccountInfo', () => {
         // Arrange
         dataDefaults = () => ({
             hasFormUpdate: false,
-            shouldShowErrorPage: false
+            shouldShowLoadErrorCard: false
         });
         cookiesSpy = jest.fn();
         httpSpy = jest.fn();
+        pushEventSpy = jest.fn();
         sutMocks = {
             $parent: {
                 $emit: jest.fn()
             },
             $http: httpSpy,
             $cookies: cookiesSpy,
-            $log: logMocks
+            $log: logMocks,
+            $gtm: {
+                pushEvent:  pushEventSpy
+            }
         };
         sutProps = {
             authToken: token,
@@ -101,7 +103,7 @@ describe('AccountInfo', () => {
 
     describe('when creating the component', () => {
         it('should register the Store Module', async () => {
-            // Arrange
+            // Arrange & Act
             wrapper = await mountAccountInfo({ storeOverride: new Vuex.Store() });
 
             // Assert
@@ -169,17 +171,20 @@ describe('AccountInfo', () => {
             );
         });
 
-        it('should set shouldShowErrorPage flag to true if an error occurs', async () => {
-            // Arrange & Act
+        it('should set shouldShowLoadErrorCard flag to true if an error occurs', async () => {
+            // Arrange
             const errorActions = {
+                ...storeActions,
                 loadConsumerDetails: jest.fn().mockImplementationOnce(() => {
                     throw new Error('some-error');
                 })
             };
+
+            // Act
             wrapper = await mountAccountInfo({ actions: errorActions });
 
             // Assert
-            expect(wrapper.vm.shouldShowErrorPage).toEqual(true);
+            expect(wrapper.vm.shouldShowLoadErrorCard).toEqual(true);
         });
 
         it('should not show the error card if no errors', async () => {
@@ -191,10 +196,10 @@ describe('AccountInfo', () => {
             expect(element.exists()).toEqual(false);
         });
 
-        it('should show the error card if shouldShowErrorPage is true', async () => {
+        it('should show the error card if shouldShowLoadErrorCard is true', async () => {
             // Arrange & Act
             wrapper = await mountAccountInfo();
-            await wrapper.setData({ shouldShowErrorPage: true });
+            await wrapper.setData({ shouldShowLoadErrorCard: true });
             const element = wrapper.find('[data-test-id="account-info-error-card"]');
 
             // Assert
@@ -204,6 +209,7 @@ describe('AccountInfo', () => {
         it('should log an error if loading preferences throws an error', async () => {
             // Arrange
             const errorActions = {
+                ...storeActions,
                 loadConsumerDetails: jest.fn().mockImplementationOnce(() => {
                     throw new Error('some-error');
                 })
@@ -223,38 +229,6 @@ describe('AccountInfo', () => {
     });
 
     describe('`methods`', () => {
-        describe('`onFormSubmit`', () => {
-            describe('form is valid', () => {
-                it('address has not changed', async () => {
-                    // Arrange
-                    wrapper = await mountAccountInfo();
-                    await wrapper.setData({ hasAddressBeenUpdated: false, hasFormUpdate: true });
-
-                    jest.spyOn(wrapper.vm, 'isFormInvalid').mockImplementation(() => false);
-
-                    // Act
-                    wrapper.vm.onFormSubmit();
-
-                    // Assert
-                    expect(wrapper.vm.accountInfoAnalyticsService.trackFormSubmission).toHaveBeenCalledWith(false);
-                });
-
-                it('address has changed', async () => {
-                    // Arrange
-                    wrapper = await mountAccountInfo();
-                    await wrapper.setData({ hasAddressBeenUpdated: true, hasFormUpdate: true });
-
-                    jest.spyOn(wrapper.vm, 'isFormInvalid').mockImplementation(() => false);
-
-                    // Act
-                    wrapper.vm.onFormSubmit();
-
-                    // Assert
-                    expect(wrapper.vm.accountInfoAnalyticsService.trackFormSubmission).toHaveBeenCalledWith(true);
-                });
-            });
-        });
-
         describe('`onEditConsumer`', () => {
             describe('when editing the form', () => {
                 it.each([
@@ -334,11 +308,28 @@ describe('AccountInfo', () => {
         });
 
         describe('onFormSubmit ::', () => {
-            it('should log an info log', async () => {
+            it('should call the save action with the correct parameters', async () => {
+                // Arrange
+                wrapper = await mountAccountInfo();
+                await wrapper.setData({ hasFormUpdate: true });
+
                 // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(storeActions.saveConsumerDetails).toHaveBeenCalledWith(expect.any(Object), {
+                    api: wrapper.vm.$data.consumerApi,
+                    authToken: token
+                });
+            });
+
+            it('should log an info log', async () => {
+                // Arrange
                 wrapper = await mountAccountInfo();
                 await wrapper.setData({ hasFormUpdate: true });
                 logMocks.info.mockClear(); // initialise has already logged info once
+
+                // Act
                 await wrapper.vm.onFormSubmit();
 
                 // Assert
@@ -349,13 +340,80 @@ describe('AccountInfo', () => {
                 );
             });
 
-            // to be added in a future pr
-            it.skip('should set shouldShowErrorPage flag to true if an error occurs', async () => {
+            it('should set shouldShowSaveErrorAlert flag to true if a save error occurs', async () => {
+                // Arrange
+                const errorActions = {
+                    ...storeActions,
+                    saveConsumerDetails: jest.fn().mockImplementationOnce(() => {
+                        throw new Error('some-error');
+                    })
+                };
+                wrapper = await mountAccountInfo({ actions: errorActions });
+                await wrapper.setData({ hasFormUpdate: true });
 
+                // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(wrapper.vm.shouldShowSaveErrorAlert).toEqual(true);
             });
 
-            // to be added in a future pr
-            it.skip('should not call the save action if no changes', async () => {
+            it('should not call the save action if no changes', async () => {
+                // Arrange
+                wrapper = await mountAccountInfo();
+                await wrapper.setData({ hasFormUpdate: false });
+
+                // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(storeActions.saveConsumerDetails).not.toHaveBeenCalledWith();
+            });
+
+            it('should set shouldShowSaveErrorAlert flag to true if a save error occurs', async () => {
+                // Arrange
+                const errorActions = {
+                    ...storeActions,
+                    saveConsumerDetails: jest.fn().mockImplementationOnce(() => {
+                        throw new Error('some-error');
+                    })
+                };
+                wrapper = await mountAccountInfo({ actions: errorActions });
+                await wrapper.setData({ hasFormUpdate: true });
+
+                // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(wrapper.vm.shouldShowSaveErrorAlert).toEqual(true);
+            });
+
+            describe('publishing analytics', () => {
+                describe('form is valid', () => {
+                    it('address has not changed', async () => {
+                        // Arrange
+                        wrapper = await mountAccountInfo();
+                        await wrapper.setData({ hasAddressBeenUpdated: false, hasFormUpdate: true });
+
+                        // Act
+                        await wrapper.vm.onFormSubmit();
+
+                        // Assert
+                        expect(pushEventSpy).toMatchSnapshot();
+                    });
+
+                    it('address has changed', async () => {
+                        // Arrange
+                        wrapper = await mountAccountInfo();
+                        await wrapper.setData({ hasAddressBeenUpdated: true, hasFormUpdate: true });
+
+                        // Act
+                        await wrapper.vm.onFormSubmit();
+
+                        // Assert
+                        expect(pushEventSpy).toMatchSnapshot();
+                    });
+                });
             });
         });
     });